+++ zeebe/protocol/src/main/java/io/camunda/zeebe/protocol/record/value/UserTaskRecordValue.java
@@ -81,5 +81,5 @@ public interface UserTaskRecordValue
    */
   long getProcessDefinitionKey();
 
-  long getPriority();
+  int getPriority();
 }
